Fixed full edit page refresh bug (See issue #39).
Changed to use of history.push in CatalogEntry and FullEdit react
elements in order to remove weird interaction with the browser go back
button.

Added an additional state variable to the CatalogEntry element in order
to keep track of the last props.part version seen by the element. This
is required in order to get the values to refresh during a rerender
after returning from the full edit page, as well as to preserve the
quick edit visual updates.